Updated openstack/openstack Project: openstack/nova-specs  67d8fe76496b337163776fe3c600218d816b3023
Blueprints and specs link has been replaced

In mitaka-template.rst file, there is a link which forwards to
a slightly deprecated site about blueprints and specs.

In the current template [1] under
"Some notes about the nova-spec and blueprint process:"

there are several points, the first one is:
 * Not all blueprints need a spec. For more information see
   http://docs.openstack.org/developer/nova/devref/kilo.blueprints.html#when-is-a-blueprint-needed

This link refers to documentation of Kilo release, and
there is information about Juno and Kilo release, which may be misleading.

This link is being replaced with [2] which is
the current version of the blueprint document.

[1] https://github.com/openstack/nova-specs/blob/master/specs/mitaka-template.rst
[2] http://docs.openstack.org/developer/nova/blueprints.html

Change-Id: I2a178966b65f628d5752c5327aeda8b625f190ba
Closes-Bug: #1582783