Updated openstack/openstack Project: openstack-infra/system-config  9ff30c68d228115350f3f6cf89772e09196c0532
Create stable order for CI results stable

One of the regressions in our pipeline fix was that Jenkins no longer
ends up on top of the results list. From a UX perspective pinning
first party results to the top of the list was very important, because
that's the first thing most people are looking for.

This reintroduces that, plus creates a stable (alphabetic) ordering of
all non first party reporters so that we get a stable reporting order,
and not an order that is merely based on who showed up first.

Change-Id: Idd38a047bc73bfe3d7ecc89291c10c0fefbfe665