Updated openstack/openstack Project: openstack/nova  87dc738763d6a7a10409e14b878f5cdd39ba805e
Make flavor-manage api call destroy with Flavor object

The flavor-manage api delete command currently need 3 separate db calls
to remove a flavor from the db. This is due to the use of flavor name to
identify the db flavor when the api command itself take flavorid as
input.

This change reworks this to only issue a single db call that now use
flavorid all way through to the db. It also gets rid of the in between
flavors module by calling destroy on a Flavor object right from the api.

Additional test changes are made to similarily bypass the flavors module
and instead call destroy() directly on the Flavor objects so the flavors
method becomes unused and thus can be removed.

Change-Id: I62bc11887283201be0bcdcfe9fa58b3f6156c754