Updated openstack/openstack Project: openstack/cinder  81453aef76ecece4e568127141cb05b99a2100b6
VMware: Update volume config during unmanage

Currently unmanage is NOP for the VMDK driver. Therefore, the
configuration which the VMDK driver persisted in the backend
object (in vCenter) will not be removed, and the backend object
may incorrectly be identified as a Cinder volume by third-party
solutions interacting with vCenter. This patch overrides the
default implementation of unmanage driver API to remove the
configuration persisted by the VMDK driver during volume
creation.

Change-Id: I13ed2b1c72d8a3cf9ff8f7f3d104e5967c3ebff1
Closes-bug: #1584683