Updated openstack/openstack Project: openstack/neutron  304b8a9ec6ce90ce49aa1063c01e39c58e65a9d7
Call ext_manager.delete_port on port removal

This patch updates OVSNeutronAgent to call ext_manager.delete_port
when the port is removed (similar to how CommonAgentLoop does).

Otherwise the QoS PortPolicyMap is not properly updated on port
removal and QoS extension will not apply policies if that port is
later reused by a VM on the same node (unless the agent is restarted).

Change-Id: I306c496985a550176d0ddc6eb33a4a4d351acb55
Closes-Bug: #1515533