Updated openstack/openstack Project: openstack/neutron  52a803caa36cfa29f4e2fa31e35c0d22cdd8d989
Add service-type extension to flavors plugin

This adds the service-type extension to the flavors plugin supported
extension aliases and enables the API tests for it. Note that the
actual logic for retrieving the listing is entirely in the extension
so no other changes were required to the flavors plugin to enable it.

The flavors plugin was chosen because the flavors plugin depends on
the service provider framework so it might as well be the one responsible
for enabling its API extension.

Partially-Implements: bp/multi-l3-backends
Change-Id: I744cc05603f4dd5dcd94a72eeb69889f5b1f38b4