Updated openstack/openstack Project: openstack/ironic-python-agent  b584bb4ddc95f3511d0cd85c4b81ec02741ce1c0
Revert "Add hardware manager interface for hardware initialization"

I would've voted -1 on the patch in question had I reviewed it, and per
standard OpenStack/Ironic procedure, I'm reverting it for re-review and
discussion.

In this case; I don't think the new method in the HWM interface is
needed, and that evaluate_hardware_support() is intended to handle the
cases handled.

This reverts commit 0962cae1da69a1a2981d5950ad741d91115dac06.

Change-Id: Ic08e44bdf116403444b257ee9f4e5b906f5eac53