Updated openstack/openstack Project: openstack/keystone  a991d9edfeac14bc2c565a779dde6e3c51158ede
Move identity.backends.sql model code to sql_model.py

This patch moves identity.backends.sql model code to sql_model.py so
that they can be shared with the shadow users sql backend.

In addition, I think it is a cleaner implementation to separate the
model from the driver.

Change-Id: I9bc44e0b9837bf5800dddd94bb36804c44581240