Updated openstack/openstack Project: openstack/ironic  246fa82890c1a9689a8d0e731f9e32911accea85
Add some docs about firmware security

We don't want users to believe that ironic handles firmware security for
them, so document that we don't, and give some advice on how to do it.

This should be the start of a larger security document that talks about
firmwares, networks, ramdisk, more?

Co-authored-by: Devananda van der Veen <devananda.vdv@gmail.com>

Change-Id: I703b5d03da610fde43dd6723edd6cdc53051c1c3