Updated openstack/openstack Project: openstack/cinder  c1c4533757d791d081a3718c1eca780c58969f75
Sheepdog:make full use of all sheepdog nodes

The sheepdog driver of cinder should use a policy to make full
use of all sheepdog nodes of a cluster instead of single node.
It can solve single point of failure and hot spot problem.

This patch uses random method. When it needs to run a dog command
or provide a location path to qemu, it gets a random node from all
sheepdog nodes of a cluster.

Change-Id: I0280b71203d99829796244afbd9a8f308b7e910a
Closes-Bug: #1560807