Updated openstack/openstack Project: openstack/horizon  0a0cafbd218b6dfc0ba6fd5bd57cf85dc7c71c79
Use 'label' rather than 'button' for buttons in Edit Image

Per the bug, it seems the styling for the Edit buttons requires that the
elements are of type 'label' and not type 'button.'  Otherwise, the buttons
look odd, particularly if you look at them in the 'Material' theme.  This
patch changes those buttons to labels.

Change-Id: I4f329513f0ec093fae84394c3094dc526e1a9b9e
Closes-Bug: 1584207