Updated openstack/openstack Project: openstack/governance  b78e7bd0e2e42d212b6e9271ceccf2c5c25688c7
Add project Vitrage to OpenStack big-tent

This is a request to add the Vitrage project[1] to the Big Tent.
Vitrage aims to be the OpenStack RCA (Root Cause Analysis) Engine for
organizing, analyzing and expanding OpenStack alarms & events, yielding
insights regarding the root cause of problems and deducing their
existence before they are directly detected.

Vitrage provides an holistic and complete view of the cloud, with
a clear visualization of the physical-to-virtual (and soon the
applicative) layers. Vitrage receives information from various
OpenStack data sources (Aodh, Nova, Cinder and Neutron), as well as
from external data sources (like Nagios, and in the future Zabbix)
and combines it in a topology graph database. This topology is used to
deduce that certain problems in the physical layer should cause problems
in the virtual layer; to trigger corresponding alarms; to modify
(currently only in Vitrage) resources states; and to reflect the causal
relationship of the alarms in the cloud.

Vitrage development started right after Mitaka Summit, where we had
meetings with other OpenStack projects (like Ceilometer/Aodh), presented
to them Vitrage ideas, verified that there was no overlap between our
projects, and got their blessing.

Vitrage follows OpenStack guidelines and “four opens” from day one.
Vitrage code and specifications are entirely open source. The project
is fully licensed under the Apache 2.0 license, and the project uses
public code reviews in gerrit. The code can be found here[2][3][4][5].
Vitrage gate runs docs, pep8, python27, python34 and tempest tests.

The #openstack-vitrage IRC channel on freenode is logged [6]. Vitrage
team holds weekly IRC meetings on OpenStack official meeting channel,
which is logged as well[7]. Openstack-dev mailing list is used to
publicly discuss Vitrage related issues. Vitrage wiki page[1] consists
of high-level and low-level design documents, presentations and demos.
OpenStack launchpad[8] is used to manage blueprints and bugs. We started
the process of Newton design discussions right after Austin Summit. We
perform these discussions in the IRC meetings, in the mailing list, and
in etherpads[9][10] created for this purpose.

Vitrage was presented in two sessions in Austin Summit[11][12], as well
as in demos in the marketplace. In addition, Aodh-Vitrage integration
was discussed in Aodh design session[13][14]. One Aodh fix was already
made to enable this integration[15]. Other integrations, with Monasca
and Congress, were discussed during the summit, and we are continuing
such discussions in the mailing list.

In addition to OpenStack community, Vitrage is also involved in OPNFV
projects. Vitrage is considered a reference implementation for OPNFV
Doctor Inspector component [16][17], as well as for OPNFV PinPoint[18]
project. Vitrage is about to be presented, together with OPNFV Doctor
and with OpenStack Congress, in OPNFV Summit in Berlin[19]. Vitrage demo
is going to be presented in Doctor POC booth.

For the past six months, since Vitrage development began, we have been
encouraging interaction and cooperation with other projects and
contributors in OpenStack. To date, a few companies and projects have
worked with us on formulating requirements as well as contributed code.
We are working hard at involving additional players in Vitrage, to help
drive its feature set so that it supports a wide range of needs from the
community.

To summarize, Vitrage fills a real need in OpenStack. Vitrage holistic
and complete view that includes all of the layers of the cloud, as well
as its alarm analysis and correlation, is currently lacking in
OpenStack. As Vitrage enhances OpenStack capabilities, raises a lot of
interest in the community, and is developed by OpenStack standards, we
believe it belongs to the big tent.

[1] https://wiki.openstack.org/wiki/Vitrage
[2] https://github.com/openstack/vitrage
[3] https://github.com/openstack/vitrage-dashboard
[4] https://github.com/openstack/python-vitrageclient
[5] https://github.com/openstack/vitrage-specs
[6] http://eavesdrop.openstack.org/irclogs/%23openstack-vitrage/
[7] http://eavesdrop.openstack.org/meetings/vitrage/
[8] https://launchpad.net/vitrage
[9] https://etherpad.openstack.org/p/vitrage-newton-planning
[10] https://etherpad.openstack.org/p/vitrage-overlapping-templates-support-design
[11] https://www.youtube.com/watch?v=9Qw5coTLgMo
[12] https://www.youtube.com/watch?v=ey68KNKXc5c
[13] https://www.openstack.org/summit/austin-2016/summit-schedule/events/9203
[14] https://etherpad.openstack.org/p/newton-telemetry-vitrage
[15] https://review.openstack.org/#/c/314683/
[16] https://wiki.opnfv.org/display/doctor/Doctor+Home
[17] https://jira.opnfv.org/browse/DOCTOR-57
[18] https://wiki.opnfv.org/display/pinpoint/Pinpoint+Home
[19] https://www.eventscribe.com/2016/OPNFV/aaSearchByDay.asp?h=Full%20Schedule&BCFO=P|G see “Failure Inspection in Doctor using Vitrage and Congress”

Change-Id: I47f7302da3d35a4620fc910c38384b4b4c71de7d