Updated openstack/openstack Project: openstack/tempest  22b9fec99e5fbd2402e91d89c536e1c59a28beed
Fix update_password tests

The current implementation uses object shallow copy to
get new credentials and clients to restore the password
after update, which is fragile and a bit complicated.

Using one client is enough. To restore the password
we clear auth to ensure a new token is requested, set the
new password temporarily, restore the password and finally
reset auth again.

Change-Id: I41f442ee89b1bd37dacb02b5ca276abd85981295