Updated openstack/openstack Project: openstack/ceilometer  3c782e4d6bae8363012531f1bd8bd20294e2f1b9
Set the time point polling starts as timestamp of samples

Many of the pollsters define the timestamp individually for each
generated sample, which should really be timestamping based on when
the data was polled and not when each sample is generated. We need to
set the timestamp of the polled data to the timestamp when the polling
starts for unity.

This commit set the time point when the polling starts as the timestamp
of samples, and make 'timestamp' argument of Sample.__init__ method
optional.

The timestamping in polling agents will be dropped in following patches.

Change-Id: I3be3f7fb8300e8853ebd35986cdf159422aa247a
Partial-Bug: #1491509
Partially-Implements: blueprint unify-timestamp-of-polled-data