Updated openstack/openstack Project: openstack/cinder  3bf84ddd14aca25c17aecfae6d9d4e5cc9e2c63f
Add hacking check to prevent assert_called_once

We've had a few patches now over the last few releases to fix invalid
mock assertions for assert_called_once. Rather than keep letting these
creep back in to the code, this adds a hacking check to watch for this
specific call and block it.

Change-Id: I3cf4df523ddba49f57dd1d4d23e95e6d62d03c9e