Updated openstack/openstack Project: openstack/cinder  7919bb781bb1de126015a41da128708b7cf8a7d3
Fix up RPC cleanup process

We don't need to assert here, because we will
throw exceptions when using TRANSPORT and
NOTIFICATION_TRANSPORT if they are None.

If NOTIFIER is None, log an exception and continue,
since it is not used here.

This still allows detection of problems like the
asserts did, but will not fail needlessly.

Change-Id: I6f0c3ce19a60339e926946b605962d4121a06d8e