Updated openstack/openstack Project: openstack/neutron-specs  a4e76357c66d3fcfacdee52cee90c402be78ad78
Add spec for push notification refactor

Adds a spec to change the method we use to get
information from the server to the agents. Rather
than the server notifying the agent to call the server,
we can just put the relevant data in the notification
itself to improve scalability and reliability.

The bulk of this spec is dealing with the message ordering
guarantee we will need to accomplish this. It also has
some work items to help improve our current pattern.

Related-Bug: #1516195
Change-Id: I3af200ad84483e6e1fe619d516ff20bc87041f7c