Updated openstack/openstack Project: openstack/tempest  a52853ed10d27224d842dd1efa01d2bf58c6eb61
Add cleanup in identity/test_users_negative.py

If the assertRaises(...) fails (e.g if an exception is not raised)
then the test will exit right away and the method call to reset the
environment is not executed.

We should use the addCleanup(...) mechanism to be sure that the cleanup
happens regardless of the status of the test.

Change-Id: I43cec136691bb1b39b01756a2a4139abf3d2d51b