Updated openstack/openstack Project: openstack/tempest  b5eb6512c85a16d1c1b6846408be4ee2f8a483c0
test_port_security_macspoofing_port: Don't assume DHCP port

DHCP port is a detail of the reference implementation.
Non agent based implementations might not provide it.
Anyway, as the purpose of this test case is to test
port-security, not DHCP, it's better to avoid assuming
the existance of DHCP port.  Use an ordinary VM instead.

Closes-Bug: #1521224
Change-Id: I84a69c6b3d81218ba7811332d936cb891bdff353