Updated openstack/openstack Project: openstack/nova  410f8af2a1d17bc67a6c0611cc7d02045c45d35e
Remove SchedulerClientTrackerTestCase

This unit test in test_resource_tracker.py was ostensibly testing to
ensure that if resources didn't change on the compute node, that the
scheduler client's update_resource_stats() method was not called (and
conversely, if the resources *did* change, that the
update_resource_stats() scheduler client call *was* called.

This was a good unit test. The problem was that this unit test already
existed entirely in the test_tracker.py test file in the
TestUpdateComputeNode test class. The test_tracker.py test classes don't
suffer from deriving from a test class base that spawns a conductor
process and init's a database (in other words, isn't a unit test). And
so goodbye duplicate SchedulerClientTrackerTestCase.

Change-Id: I4905883dc4372fff413f009ec9110985998542bd