Updated openstack/openstack Project: openstack/oslo.config  e7b6cef9db8cf8902f467f4ba25fb7dc2d37168a
Incorrect group name when deprecated_group is not specified

When an opt is defined with a deprecated_name alias but not with
deprecated_group, the generated conf file assumes 'DEFAULT' section.

i.e say we do :-
opt = cfg.StrOpt('cinder_os_region_name', deprecated_name='os_region_name',
                  help=_('..'))
conf.register_opt([opt], group='glance_store')

we get the following in the generated conf file :-
 Deprecated group/name - [DEFAULT]/os_region_name
 cinder_os_region_name = <None>

Instead of,
 Deprecated group/name - [glance_store]/os_region_name
 cinder_os_region_name = <None>

On contrary, when you register such an opt, the deprecated_name is
associated against the registered group. Thus, the generated conf
is not in sync.

We found this issue when generating glance-api conf's but this
could be affecting other projects as well.

Closes-Bug: #1554657
Co-Authored-By: Tan Lin<lin.tan@intel.com>
Change-Id: I9577abec87b20fc3d74fb71e327e32d7def8ac5d