Updated openstack/openstack Project: openstack-infra/project-config  eda18bc7515c4c8a0969dac7efbf5ca3f10d89ac
Create stackviz element and script in nodepool

In order to get Stackviz running against tempest gate run data, it
will need to be pulled in its pre-built form (i.e. with npm deps
installed) from the nodepool images. This change adds a couple steps
to build Stackviz so that it is in a ready state for retrieval and use
on the logs server.

A future devstack-gate patch will pull the stackviz directory from the
nodepool images onto the logs server. This nodepool patch is required
first in order to prepare stackviz for download. Installing the npm
modules takes a considerable amount of time (~5-10 mins), so it is
best that this work is done on the images before gate jobs are run.

Depends-On: I383e5bf99f6e9e2c7d5fa5c894ef573fa06facd7
Change-Id: I75fdad6584ee237dd07e25148a27ff4df3a5efe1