Updated openstack/openstack Project: openstack-infra/project-config  553719831e71a9b67c339ce5248ae417331e11c4
Promote OSA infrastructure CentOS jobs to non-voting

This patch promotes the infrastructure-related jobs on CentOS 7
in OpenStack-Ansible from experimental to non-voting. The
OpenStack roles are left as experimental for now as they rely
on the infrastructure roles for testing.

There is also a small correction for the os_aodh repo where the
Xenial experimental job was mistakenly moved. This patch also
puts it back in the experimental queue.

Change-Id: Ic56c29d73cd3725d8e29e58be0403f9e0e1e8cf3