Updated openstack/openstack Project: openstack/ironic-python-agent  ced9a754dd521df318f7a60b38e876619110b735
Use constraints for all the things

This configures tox to use upper-constraints for everything, except
environments used for post jobs (as constraints aren't supported there
yet).

This also doesn't enable constraints for our image builds, as it isn't
possible to pass an environment variable into the Dockerfile to allow
folks to choose their own constraints file (and we don't want to force
people to rely on git.o.o for building images).

Change-Id: I285855c6386f59e9dd352473aa28463109818c90