Updated openstack/openstack Project: openstack/keystone  79952ffbd4c1ffd2dab32c04581b3a7f71a05e28
remove deprecated revoke_by_expiration function

revoke_by_expiration is only useful if a token does not have an
audit_id or audit_chain_id. tokens always have an audit ID, so
this function seems redundant. it was also deprecated in the
J release, with no timeline for its removal

Change-Id: Ieb92a70ab782fa8ceb59dc807ea8647587be9e2b
bp: removed-as-of-newton