Updated openstack/openstack Project: openstack/heat  f3033f65ef3cffb7ab55ab462d63e061b72e4ece
Convergence: show parameters while deleting stack

In case of convergence, delete is an update of stack with empty
template. The issue with it is that the stack-show will not show the
parameters used to create/update the stack. The fix is to retain the
environment from the last template in the empty template and then update
the stack with that template so that the stack-show will show the
parameters properly.

Change-Id: Ia45ad5d024456cd583c282a42b26c9528512e73c
Closes-Bug: #1583673