Updated openstack/openstack Project: openstack/keystone  faa79c8e183a6d8383c8e34ca737aa20fc3cf693
Add API Change Tutorial

This change adds a tutorial for making an API change. It describes from
the specification proposal to making real code changes in keystone.

Its goal is to help new contributors to get familiar with keystone code.
It simulates the addition of a 'description' field to role entities.

Change-Id: Ie6f302939f43e78f07183abf4bc5aadb6b50ef1c