Updated openstack/openstack Project: openstack-infra/project-config  abacb8faed3afdc5ada4c769b41f3a285499840b
Remove integrated jobs from neutron-dynamic-routing project config

The neutron integrated jobs (gate-tempest-dsvm-neutron-full and
gate-grenade-dsvm-neutron) are not configured to run the dynamic
routing extension, so running these jobs is not only wasteful but
slow down pace of develoment.

Since there is not even a remote chance that commits against this
project can break the integrated jobs, they can be safely removed.

Change-Id: I77401df645451c9ae218bd8b9855cc2725de9756