Updated openstack/openstack Project: openstack-infra/storyboard  cd6447ffb5df37be664676e7992f4f56b77f71ab
Avoid false positives when filtering projects by subscriber

At the moment, filtering projects by subscriber also returns the
projects with the same ID as any project groups that the subscriber
is also subscribed to.

This commit fixes that by filtering for an exact match of 'project',
rather than anything which includes the string 'project'.

Change-Id: I3a9f863e8d8dd18f4b5e3fe9c9fe44ee3c72457c