Updated openstack/openstack Project: openstack/ironic-python-agent  f412b931cb30f095c039fcc82696c83c954a837b
Remove links to github for OpenStack things

Github is not the official reference for OpenStack code repositories,
and as such should not be used in documentation.

Update links that are meant to be links to a project as a whole with
links to the project's documentation.

Update links directly to code to git.openstack.org.

This change follows Iee09b6b86628e46dc5e9436bd0f98a40c50244b7

Change-Id: I6c0ba75c1ba04916413847cce56441ff4670bd62