Updated openstack/openstack Project: openstack/heat  c8b2f2bcaa6036ff010ce953ba0fab118bbae9d7
Don't support floating ip if only specify 'network'

If only specify 'network' and 'floating_ip' for server nics,
the floating ip isn't be associated to server ports.
Due the ports are created out of heat, and heat doesn't know which
port should the floating ip to associated with. This change adds
validation of the case.

Change-Id: I1ee61d242d8ec9d756bd38f6b03c0751f740fd63
Closes-Bug: 1571992