Updated openstack/openstack Project: openstack/neutron  1ce8ce9546479c0ce6055c0c205a8885699e3051
tests: guarantee unique networks in get_list_of_random_networks

The function generates a list of random networks, but so far did not
guarantee their uniqueness. In rare cases, it breaks
test_subnetpool_prefixes unit test in gate that relies on those networks
to be unique.

Now, the function will regenerate the result several times if duplicates
are detected.

Change-Id: Ia300e36ad7a25dd156ac8b5f27dcdeeaa5339ab8
Closes-Bug: #1585941